修正编译错误 消息可以通过protobuf的反射自动查找消息号，虽然会有点损耗，总比写错了消息号来的好

+++ server-res/res/db/db_proto/gamedb.pb.sql
@@ -136,8 +136,8 @@ CREATE TABLE `tbld_pet` (
 DROP TABLE IF EXISTS `tbld_player`;
 CREATE TABLE `tbld_player` (
   `id` bigint(20) UNSIGNED  NOT NULL AUTO_INCREMENT COMMENT '',
-  `worldid` int(11) UNSIGNED  NOT NULL COMMENT '服务器编号',
-  `oriworldid` int(11) UNSIGNED  NOT NULL COMMENT '创建角色时的服务器编号',
+  `worldid` smallint(11) UNSIGNED  NOT NULL COMMENT '服务器编号',
+  `oriworldid` smallint(11) UNSIGNED  NOT NULL COMMENT '创建角色时的服务器编号',
   `openid` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '玩家账号',
   `name` varchar(32) CHARACTER SET utf8 NOT NULL COMMENT '玩家名字',
   `prof` int(11) UNSIGNED  NOT NULL COMMENT '玩家职业',
