+++ server-res/res/db/db_proto/globaldb.pb.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS `tbld_dbinfo`;
 CREATE TABLE `tbld_dbinfo` (
   `worldid` int(11) UNSIGNED  NOT NULL COMMENT '游戏服编号',
   `db_ip` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '游戏数据mysql地址',
-  `db_port` int(11) UNSIGNED  NOT NULL COMMENT '游戏数据mysql端口',
+  `db_port` smallint(11) UNSIGNED  NOT NULL COMMENT '游戏数据mysql端口',
   `db_name` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '游戏数据mysql表名',
   `db_user` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '游戏数据mysql用户',
   `db_passwd` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '游戏数据mysql密码',
@@ -36,26 +36,26 @@ CREATE TABLE `tbld_gmlist` (
 
 DROP TABLE IF EXISTS `tbld_serverlist`;
 CREATE TABLE `tbld_serverlist` (
-  `worldid` int(11) UNSIGNED  NOT NULL COMMENT '游戏服编号',
+  `worldid` smallint(11) UNSIGNED  NOT NULL COMMENT '游戏服编号',
   `server_name` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '服务器名',
   `open_time` bigint(20) UNSIGNED  NOT NULL COMMENT '开服时间戳',
   `open_status` int(11) UNSIGNED  NOT NULL COMMENT '开放状态',
   `publish_addr` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '入口地址',
-  `mergeto` int(11) UNSIGNED  NOT NULL COMMENT '已经被合服到哪个服了',
+  `mergeto` smallint(11) UNSIGNED  NOT NULL COMMENT '已经被合服到哪个服了',
   PRIMARY KEY (`worldid`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 
 DROP TABLE IF EXISTS `tbld_servicedetail`;
 CREATE TABLE `tbld_servicedetail` (
-  `worldid` int(11) UNSIGNED  NOT NULL COMMENT '游戏服编号',
-  `serviceid` int(11) UNSIGNED  NOT NULL COMMENT '服务编号',
+  `worldid` smallint(11) UNSIGNED  NOT NULL COMMENT '游戏服编号',
+  `serviceid` smallint(11) UNSIGNED  NOT NULL COMMENT '服务编号',
   `service_type` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT 'lib库名',
   `route_addr` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '内网通讯地址',
-  `route_port` int(11) UNSIGNED  NOT NULL COMMENT '内网通讯端口',
+  `route_port` smallint(11) UNSIGNED  NOT NULL COMMENT '内网通讯端口',
   `publish_addr` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '外网地址',
-  `publish_port` int(11) UNSIGNED  NOT NULL COMMENT '外网端口',
-  `debug_port` int(11) UNSIGNED  NOT NULL COMMENT 'debug端口',
+  `publish_port` smallint(11) UNSIGNED  NOT NULL COMMENT '外网端口',
+  `debug_port` smallint(11) UNSIGNED  NOT NULL COMMENT 'debug端口',
   `bind_addr` varchar(255) CHARACTER SET utf8 NOT NULL COMMENT '监听绑定地址',
   PRIMARY KEY (`worldid`,`serviceid`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
